refactor: xmodule/word_cloud_module.py -> xmodule/word_cloud_block.py
