+++ xmodule/static_content.py
@@ -29,7 +29,7 @@ from xmodule.poll_block import PollBlock
 from xmodule.seq_block import SequenceBlock
 from xmodule.split_test_module import SplitTestBlock
 from xmodule.template_block import CustomTagBlock
-from xmodule.word_cloud_module import WordCloudBlock
+from xmodule.word_cloud_block import WordCloudBlock
 from xmodule.x_module import HTMLSnippet
 
 LOG = logging.getLogger(__name__)
